Added remind me button for testing push notifications
