+++ package.json
@@ -11,6 +11,7 @@
     "precommit": "lint-staged"
   },
   "dependencies": {
+    "@react-native-async-storage/async-storage": "~1.15.0",
     "expo": "~42.0.1",
     "expo-constants": "^12.1.3",
     "expo-notifications": "^0.13.3",
@@ -22,8 +23,7 @@
     "react-native-modal-datetime-picker": "^12.0.0",
     "react-native-select-dropdown": "^1.2.0",
     "react-native-web": "~0.13.12",
-    "react-native-weekly-calendar": "^0.2.0",
-    "@react-native-async-storage/async-storage": "~1.15.0"
+    "react-native-weekly-calendar": "^0.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.9.0",
